+++ functions.h
@@ -32,7 +32,7 @@ float circumference(int radi);
 float diagonal_rectangle(int len, int wid);
 float exterior_angle(int angles);
 float interior_angle(int angles);
-float area_regular_polygon(int side, int angles);
+float area_regular_polygon(float side, float angles);
 int prompt(int request);
 
 #endif
